main AudioManager
made AudioManager (and AudioManagerSound) which just controls the main BGM tho could also be for other sounds. Also made it and GameManager into prefabs and they should both persist across scenes. The BGM also goes to 20% volume while paused. Right now the test audio is in the scripts folder, it's there to be obviously out of place so that we're sure to delete it once we get real audios