Enemy Anti Follow Player
When the enemy starts to attack the player, it has a max range that it can move away from its original position where the player was. After it has reached that limit, the enemy will ignore the player for N amount of seconds.

Changed the EnemyDetector function to return a GameObject instead of a postion to get info about the entity